Don't create a new user entry just for visiting the homepage.
When a user visits the main snippet-server homepage, we show all the
snippets for that user.  But we don't actually need a user record at
this point.  Don't create one, since the user may just be browsing and
may not want to be part of the snippet system.  Instead, create the
user object the first time the user saves a new snippet.  (Or goes to
visit the 'user settings' page.)