Modify hipchat to read the file from the filesystem rather than the datastore.
This is based on tom's review feedback.  Using the datastore is a
kludge.  We should do the same thing we do for the khan app, and just
have a 'secret' file that's not checked in but gets uploaded via
appcfg.

As part of this, I reverted connection.py to be the same as in
upstream python-hipchat.  But I still didn't want to use configobj,
which is overkill for this, so wrote a one-line config-parser
instead.  (It's considerably more brittle, which is fine for us.)