Move the reminder-emails to sunday to get around our 100-emails-per-day quota.
The 'free' quota is 100 emails a day.  So we double our email capacity
if we send reminders on sunday and 'time to view' on monday, or
reminders on monday and 'time to view' on tuesday.  For now I've
decided to do it the first way around.  Could be convinced
otherwise...