Require users to register (verify settings) before adding new snippets.
Before this change, it was too easy for folks to go to
weekly-snippets.appspot.com while logged in as the wrong user
(foo@gmail.com instead of foo@khanacademy.org, say), and enter
snippets under an email they didn't mean to.

With this change, when we see a new email, we force people to confirm
their user settings before they can enter snippets.  For an existing
user, this will serve as a heads up that maybe they're logged in as
the wrong user.  A new user will just have an extra click step before
starting up.

This has the extra benefit that a new user has the chance of filling
in the "category" field properly.