Fix _TODAY to actually return the current date (when not mocked out).
Summary:
I added _TODAY as a substitute for datetime.date() to ease testing,
but it had the side effect, which I didn't appreciate at the time,
that _TODAY was set at script-load-time and then never changed.  So
when deciding what week of snippets to show by default, for instance,
we would make the decision based on what time it was when the script
started, rather than the current time.  Oops!

By changing _TODAY to _TODAY_FN, which is evaulated every time we need
to know the current date, I avoid this problem.

Reviewers: benkomalo, tom

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D185