Add some db.get() calls in preparation for moving to HRD.
Summary:
The master/slave datastore is deprecated, so we're moving snippets
over to the high replication datastore.  We want to keep our old,
synchronous semantics, though, so we just put in some key-gets to
force the HRD to be immediately consistent.

The db model here is simple, so the changes are too (thankfully).

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D686