Change app.yaml to point to the new app-id for weekly-snippets.
Summary:
We're moving weekly-snippets to the High Replication Datastore, since
the old master/slave datastore is deprecated.  The migration process
requires creating a new app, moving all the data to the new app, and
then making the old name an alias to the new name.  This is (near) the
beginning of that process.

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D691