change todo in snippets.py
Summary:
aside from very small changes to user_snippets.html, adds in javascript/ and snippets.js in particular which is responsible for the lion's share of 'ui' changes. Things you can look forward to:
* a save button which does not reload the page.
* an undo button which restores your snippet to its last saved state.
* tests for POST methods in test_snippets.py

Test Plan:
update two snippets at once. Notice that the second snippet remained unaffected and that the save/undo buttons are disabled for untouched snippets. saving a snippet will cause the buttons to be disabled (although richer feedback is not provided). undoing a recently saved snippet will cause the snippet to revert to its state as of last undo. attempting to close, reload the snippet window with unsaved edits prompts the user to make sure if they really want to close the window.

test_snippets.py has a new PostTestCase which tests the behavior of the POST endpoint, its status codes, its responses, attempting to update as a different user, and testing that modifying a single entry via POST does not affect others.

Reviewers: csilvers

Reviewed By: csilvers

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D812