Note (in red!) when a snippet is due.
Summary:
This is in preparation for maybe putting next week's snippets up
earlier.  The red marking will draw people to the proper box to fill
in.

I also say "for the week starting XXX" rather than "for the week of
XXX", to make it clearer what the week is.  We could do the whole
range, but that's a bit more logic (especially around new years!) so I
don't bother yet.

Finally, I fixed up the tests to all pass again.  All the login tests
had started failing when I changed the code to automatically redirect
to the login screen rather than making the user click on a 'log in'
button.

Test Plan: ./snippets_test.py

Reviewers: marcos, chris

Reviewed By: chris

Subscribers: khirasaki

Differential Revision: http://phabricator.khanacademy.org/D7417