Rewrite snippets.js
Summary:
As we've added more functionality to snippets.js, it's become a bit of a soup. After a non-obvious-but-should-have-been-obvious bug with event targets landed yesterday, it seemed time to restructure some of the copypastas in snippets.js. Hopefully this makes adding new features to the User Snippets page a bit less intimidating going forward :)

Speaking of which, this adds a feature! As mentioned in D24254,

 > Known issues: toggling private / markdown state doesn't register a snippet as "dirty". This has never worked, as far as I can tell.

Now it does! Checkbox state is recorded, so they now contribute to a snippet's "dirty" status. They also reset when "Clear Changes" is clicked.

I still think that a React rewrite would be awesome, and if this file goes away //tomorrow// I'd be happy. But for now, hopefully it's a step in the right direction.

Depends on D24254.

Test Plan:
- Clicked around and typed a bunch with the console open,
- would appreciate a second pair of eyes on it if you've got a few minutes.

Reviewers: michelle

Reviewed By: michelle

Subscribers: anju

Differential Revision: https://phabricator.khanacademy.org/D24257