Move hipchat/slack integration from config files to datastore values.
Summary:
Now you can use /admin/settings to configure hipchat and slack
integration.  This makes it much easier to administer them without
needing to redeploy the app, and it also simplifies the app logic for
when to send to hipchat/slack and when not to.

Fix the name of the chat cronjob.

It's no longer hipchat-specific.

Minor refactors based on review.

1) Move the slack-formatting out of command_usage() and into the
caller, to simplify the other use of command_usage() in
app-settings.html.

2) Use stuff after the last @ as the domain more consistently

3) Fix a few tests I accidentally broke after a last-minute function
rename.

Test Plan:
make check

(none)

make check

Reviewers: benjaminpollack, mroth

Reviewed By: mroth

Subscribers: riley, michelle

Differential Revision: https://phabricator.khanacademy.org/D24375