Use ajax to send hide/delete requests.
Summary:
Added some basic error-checking, and made it work on firefox.

I don't know the best way to turn an event into an element, but I
found a way that works on two different browsers at least.  Hopefully
that's enough...

The error checking is //very// basic, fwiw.

I also made the class names more specific, so we don't match other
things accidentally.

Test Plan:
Ran 'make serve' and visited
   http://localhost:8080/admin/manage_users
as a logged-in-as-admin user (which I had already created).

I then clicked 'hide' for my one user and saw it change to 'hiding...'
and then 'unhide', I reloaded the page and saw it say 'unhide',
meaning the setting had stuck.  I then clicked on 'unhide' and saw it
change back.

I then clicked on 'delete' and saw it change to 'deleting...' and then
'deleted', and reloaded the page and saw that that user was gone.

Fixes #15

(same)

Reviewers: riley, michelle

Reviewed By: michelle

Subscribers: mroth, benjaminpollack

Differential Revision: https://phabricator.khanacademy.org/D24481