Move the hostname from being hard-coded to being a app setting.
Summary:
This is the last piece (I know of) where we are hard-coding our
dependency on our app and where it lives.

Test Plan:
make check

Also ran
   make serve
and visited
   localhost:8080
as a logged-in admin user.  I made sure there were no existing
datastore entities (using localhost:8000) and got sent to the
app-settings screen.  It defaulted to 'http://localhost:8080'
for the hostname setting, which is what it should be.

Reviewers: riley, michelle

Reviewed By: michelle

Subscribers: mroth, benjaminpollack

Differential Revision: https://phabricator.khanacademy.org/D24556