Add an actual README, in preparation for opensourcing.
Summary:
Add an "administration" section of the README.

This made me realize there was another KA-specific setting still
lurking in our codebase, in a yaml file, so I got rid of that too.

While in the area, I also got rid of the obsolete --oauth2 appcfg
flag.

Test Plan:
Will make this a pull request, and we can see how it renders!

make deploy APP=weekly-snippet-hrd

Reviewers: riley, michelle

Reviewed By: riley, michelle

Subscribers: mroth, benjaminpollack

Differential Revision: https://phabricator.khanacademy.org/D24682