Authenticate slack users by their slack ID, don't trust emails
The email field in a slack user profile isn't necessarily trustworthy -
some workspaces allow users to edit it themselves.  It also doesn't
necessarily match the user's Google account email.

This change adds a `slack_id` attribute to User entities in the
datastore, and the bot prompts users to manually configure this when
they first use a /snippets command in slack.

We can improve the onboarding UX in the future - I'm sure there is a way
of establishing Google logins from a slack app without copy-and-paste.